adding app sub folder structure
