+++ App.js
@@ -1,19 +1,5 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import {AppRegistry} from 'react-native';
+import app from './app/index';
 
-export default function App() {
-  return (
-    <View style={styles.container}>
-      <Text>Open up App.js to start working on your app!</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+AppRegistry.registerComponent('HFIT', () => app);
+export default app;
